Build and maintain LCurve in VspSurf.
This uses tessellation adaptation as a surrogate for integration
adaptation.  None of the adaptive integrators in Code-Eli will return the
u parameter.  It seemed easier to adapt this than to rewrite those.

The tessellation adaptation algorithm actually calculates the distance
between points as it goes along, so overall this is fairly wasteful.
Fortunately, it shouldn't matter.