Introduce PGFace::FindPrevNode and use in GetNodes.
This more sophisticated treatment is needed because edge loops can
double-back on themselves.  If you started trying to determine a loop
orientation from a double-back node, you would not be able to tell
anything.

This moves us off of a double-back edge, etermines orientation, and then
backs us up to the input edge.