Silence Windows error message when ccx/cgx not found on load.
These messages:

'ccx' is not recognized as an internal or external command,
operable program or batch file.
'cgx' is not recognized as an internal or external command,
operable program or batch file.

Are issued to stderr when ccx or cgx do not exist.  This change
redirects stderr to nul in this case to silence the message without
changing the result of redirecting stdout to temp.txt.