OpenVSP 3.35.0
This release is a little modest, but there is still enough here to
warrant a release before the Workshop.  I'm looking forward to seeing
some of you in two weeks -- unfortunately, if you aren't registered,
the registration is full.

The most user-visible change here is an update to the Mass Properties
capabilities.  Before, only a Blank component could be assigned a
point-mass value (and only a mass, no inertias).  With this update,
any component can be assigned mass properties -- including mass, local cg,
and inertias for the component.  (What do you call a point-mass with
non-zero inertia?)

These component mass properties will be included with the overall mass
properties calculation along with the volume-based and wetted area-based
calculations already performed.

Existing files with point-mass assigned to Blanks will be converted to the
new system -- this is not a backwards compatible change.

Blank components can now have symmetry (like all other components).  That
way, you can easily add a symmetrical copy of a Blank's mass properties.

The under-the-hood mass properties calculations have had some small changes
that should slightly increase accuracy (in ways that nobody will really
care about, but that make me happy.)

Most of the rest of the changes are very much behind-the-scenes.

Some API calls have been added to enable creating/deleting User Parms
from the API.

OpenVSP's PMARC 12 file writing was slightly improved and some Python
scripts for working with PMARC input/output files are now included.

The ability to read an existing VSPAERO run from disk was added to the
Analysis Manager -- exposing it to the API.

The biggest under the hood change should be invisible to users (except
for hopefully fewer crashes).  I've begun replacing a fundamental
library with an alternative.  The old library has a great reputation
(and we've used it forever), but when things go bad, it does not fail
gracefully -- it crashes the entire program.  This new library seems
robust and reliable so far -- but I'm hoping that if there are problems,
it will fail much more gracefully.

You will mainly see this difference in CompGeom and friends.  If this
goes well, we will later also switch to this new library in CFDMesh and
FEAMesh.  It isn't quite ready for that yet.

There were a few improvements contributed from the community.

Thanks Cibin Joseph for a bug fix when writing Degen Geom files.

Thanks to Jeff Allen for making it possible to delete multiple parameter
links at one time -- making the Add-All buttons much more useful.

All that and a few bug fixes here and there.

If you can't make Oshkosh this week, download a new OpenVSP and make
yourself an airplane.

Features:
 - Specified mass properties extended to all components
 - Point masses can now have inertia specified
 - Blanks can now be assigned symmetry properties
 - User Parm management added to API
 - pyPMARC added to python package

Libraries:
 - Use Delabella CDT library for CompGeom & friends

Bug fixes:
 - Fix missing fields in DegenGeom file.
 - Fix halfmesh bug for VLM with vspgeom files to vspaero.  Thanks Tim.