Disable corner chain treatment.
This corner chain treatment was added to help with extended rounded
wingtips.  It is unfortunately wreaking havoc with CFD and FEA Mesh
in general.

Most likely because the added chains need to be intersected with other
chains (and those added by itself).

Debugging this will take some time.  Disabling it in the interim will
help people be productive in the nearer term.