Change how disabling ALL Geom Screens is handled
Previously, this was a special case flag when the flag was checked,
this required allocating the flag array to larger than the number of
screens (to hold the ALL flag).

Instead, it now loops over all the flags and sets them as desired.
That way, a user can set all the flags one way (say to off) and then
come back through and un-set just the selected ones.