Change advanced link variable renaming to explicit dialog window.
New name was pulled from input field and event triggered when field was
chagned.  This was a functional overload of the name field that could
cause confusion.

Now an explicit button is pressed to request a rename.  This presents
a dialog window with a new name field and option boxes.