Update TMesh::SwapEdge to maintain all edge pointers correctly
Historically, Edges are not critical for TMesh operation.  They are
generated on an as-needed basis.  This means that many operations allowed
them to become invalid.

This works in some situations, but not when an ongoing operation that
would invalidate them also needs them to be correct.

This change updates SwapEdge to keep all the pointers valid as edges are
swapped.