Add PGMesh::GetOtherNodes
Given a list of nodes in a polygon, return all the polygon's nodes not
in the supplied list.

Given the two nodes of an edge, it can be used to return the 3rd node of
a triangle, or the two nodes of a quad that are not on that edge.