Use DeleteFile instead of system( del ) on Windows
del doesn't accept / as a path separator, even though it is returned by
the Windows API.  Even a call to preferred_separator() won't change it.

This will make the temp file actually be deleted and will also silence
the 'Invalid switch - "Users".' warning that shows up in the command
window.