VSPAERO CMake copying fixes for Windows.
Specifying the path terminated with a / allows the file name to be re-used,
this allows .exe to appear on  Windows, but not on Linux and Mac.

Using $<CONFIG> allows multi-config form to be used where appropriate.
This has some risk of breaking other single-config platforms.  May need
to be made conditional.

Together, these changes allow vspaero & co to be copied to the right
places to be found via the API and also regular GUI Windows use.