Make PopupMsg work from API, also change to const string & from char*
API would create temporary variable that would go out of scope before
the fl_message could be displayed.  This would cause a segfault or would
show a corrupted message.

Allocating memory, making a copy, and then deleting the copy after use
makes sure the message stays in scope for the duration of use.