Improve constiness of Geom::GetSurfPtr & co.  Don't return const ref.
Returning a const reference is fast -- unless you assign it to a local
variable (which we did every time).  This results in a copy and is no
faster than other approaches that require a copy.

This was frequently done in places like UpdateParmLimits() which
could introduce noticible interactive delays (UI latency).