Re-write ConnectBorderEdges to work on Nodes and use NanoFlann
In certain FEAMesh cases, CBE would be the dominant use of CPU time.
It used a naive algorithm to match border node points on different meshes
and to force them to be coincident.  Theoretically, this should not need
to be done as they should be coincident already.

This change uses NanoFlann - a fast nearest neighbors search to replace
the exhaustive search.

This also changes to a node-focused approach instead of an edge-focused
approach.  This change should allow merging to occur where more than two
edges meet (such as thick-thin intersections).

The new algorithm is almost immeasurable in time.