Attempt to fix FEA visualization behavior.
Optimizations made earlier to better follow DrawObj::m_GeomChanged
had the effect of making some DrawObj not show up when they should.

This is not the best fix.

Structures uses hide/shown flags to determine whether or not to
LoadDrawObj.  Instead, all DO should always be Loaded.

The m_Shown flag should be used to manipulate whether it appears or not.

Some of these routines would still need a MarkDOChanged() type routine
to handle when a pre-existing DO (that should not be recalculated) is
added to a OD set.

Possible to also fix this on the other end of the LDO process -- A
new DO should always be deep loaded, independent of m_GeomChanged.