OpenVSP 3.38.0
Coming just a month since the last release, you won't be surprised that
this version is somewhat small overall.  It includes some more fixes
that I thought about releasing as 3.37.3.  However there is also a
broad variety of quality of life improvements that this release
deserves a .0 update.

In the Geom Browser, the Show Only, Show, and NoShow buttons now work
like the other things (like Hidden, Shade, Feature lines) with respect
to applying changes recursively.  I.e. now they only recurse a collapsed
tree.  When you Show or NoShow a parent entity on an expanded tree,
the visibility change only affects the selected entities.

The number of Sets is now user-adjustable.  Go to the Set Editor,
click the arrows at the top of the window.

Measure (probe, ruler, protractor) now works for non-surface Geoms.  This
is mostly for blank and hinge geoms, but works in a rudimentary way
for Mesh and HumanGeom (it always measures from their component origin).
So users who use Blanks extensively as intermediate coordinate systems
can now use them for all their Measure needs.

Several improvements around advanced links.  Unique variable names are
now enforced when renaming variables.  The variable renaming buttons
were re-worded for clarity.  Multi-line error messages are now displayed
when links break.

The full-path filename at the bottom of the screen is now displayed
in a larger font for better visibility on high resolution screens.
The filename (without a path) is now displayed in the title bar
at the top of the screen.

The integrated flowpath engine extension has a pair of improvements.
First, streamtubes are now extended to a constant X value -- even
when the start of the streamtube is not at a constant X.  There is also
a new option to auto-determine the extension distance based on the
overall dimensions of a selected Set.

Height/width/area/aspect sliders are now displayed for CEdit type
XSec curves -- even when the detailed editing window is not displayed.
This should allow for quick adjustments without having to get deep
every time.

Better patch names are assigned to wing patches when exporting STEP
and IGES files (Thanks Jeff V.).  This should make it easier to identify
wing trailing edge, upper, and lower surfaces in meshing tools.

There is also now an option to merge the lower and upper trailing edge
patches together (and leading edge too).  Some meshing tools did not
appreciate that we output split TE patches -- so now we'll merge them.

As usual, there are a few bugfixes too -- mostly related to graphics
glitches that were introduced in 3.37.0.

I don't think there are any high risk changes in this -- and hopefully
the graphics glitches introduced in 3.37.0 are calming down.  Everyone
should update.


Features:
 - Show Only, Show, NoShow buttons now recurse like everything else.
 - User adjustable number of Sets
 - Measure works with non-surface Geoms (blank and hinge)
 - Unique variable names enforced when renaming advanced link vars
 - Re-worded variable renaming buttons
 - Output multi-line error messages when links are broken
 - Bigger filename at screen bottom
 - Filename in main title bar
 - Integrated flowpath modeling extension now extends to constant X
 - Integrated flowpath modeling extension distance auto determined
 - Height/width/area/aspect sliders for CEdit XSecCurve now displayed
 - Better patch names for unintersected STEP/IGES surface export
 - TE surface merging for STEP/IGES export

Library Updates:
 - Update to latest FLTK dev as they approach 1.4.0

Bug Fixes:
 - Fix invisible subsurfaces
 - Fix XSecSurf highlights that won't go away
 - Expand ranges for AR and Span - should fix some odd wings seen in Group
 - Fix some odd scattered indexing bugs
 - Fix some problems with CFDMesh sourcing