OpenVSP 3.39.0
Coming less than a month after the prior release, I thought about holding
off on this release until there was more to go with it.  After all, this
release really consists of just one feature and a handful of bug fixes.

However, when I've shown the new feature to users, they want it released
yesterday -- or sooner if possible.  Who am I to argue?

OpenVSP (and its predecessors) have long had the ability to place a 2D
image as a background to be used as a guide when setting up a model.
This version takes that idea to the next level.  Instead of the image
appearing as a static background, it is placed in the model such that it
can pan and zoom appropriately.  Likewise, there is no limit to the
number of reference images that can be used.

Describing this feature doesn't succumb to the written word -- so look
for a video from me or Brandon in the near future.

Aside from that, there are a couple of bug fixes mixed in.  We're still
working through the aftermath of the DrawObj m_GeomChanged enforcement
from a few versions back.  A CFDMesh tolerance has been tightened
so cusped airfoils won't get smooshed.  And some continued build system
improvements to help with portability to more diverse platforms.

We've also updated (again) to the latest FLTK 1.4.0-pre release.  The
FLTK dev team is on a big push to 1.4 and I'm trying to make sure there
aren't any surprises when it goes final.

Features:
 - 3D Background images

Library Updates:
 - Update to latest FLTK dev as they approach 1.4.0
 - Build LibXml2 with -PIC
 - Update GLEW build integration to match modern GLEW

Bug Fixes:
 - Move Documentation build to MacOS from Ubuntu to get latest Swig
 - Drop MacOS-11 build, move to MacOS-12
 - Use gcc-11 on MacOS for ADEPT and OMP builds
 - Fix MeshGeom DrawObj issue
 - Tighten tolerance in MatchBorderNodes - Thanks Andy.