Use SealDoubleBackNodes instead of ColinearEdgeLoops
Hopefully this is more robust and faster.

The search process should be faster because every node gets checked
without recursion.  The EdgeLoop detection performed a potentially
expensive recursion starting from every edge.  This would sometimes
go into infinite loops that have not yet been fixed.

The node sealing process still involves recursion, but it is only
walking along a crack as it is sealed -- it is not performing an
exhaustive search for closed loops.