Fix problem with subsurface line perfectly on top of mesh line.
Extract uvec and vvec from TMesh::m_UWPnts if available.

Identify constant U or constant V subsurface line segments.  If they are
within tolerance of a value in uvec or vvec (respectively), skip
adding them.

The geometry will be identical -- the line is already in the mesh.
Intersections should not be affected.  Classification (SubTagTris) will
not change.

This amounts to -- don't cut what is already cut.