Don't embed resources with pandoc.
This was causing image files to be embedded in the HTML.

Since the css is no longer embedded, it now needs to be copied to wherever
the html will be.  It does not need to be manually copied to the help
build directory because the RESOURCE_FILES directive to the pandoc
CMake command takes care of that.