Make MakeU/VTess obey n_ref.  Don't pass args as const refs.
Most of the args converted to pass by copy were simple scalar args, so
there is no significant performance difference.  The one that may
stand out is passing vector<int>num_u.

A copy had to be allocated anyway, so it seems reasonable to let the
compiler take care of the copy.  It also simlifies the diff a bit.