Use PGFace::m_ID as a flag to ensure faces merged once and only once
Faces that didn't get merged still need to have their jref / kref value
halved each iteration.  So we needed to be able to identify faces that
had not been merged in a regular pass.

Other faces were being merged twice in a regular pass, so we needed to
be able identify faces that had already been merged.

We did this by overloading m_ID.  First, we ResetFaceNumbers() to make
sure all face numbers are positive.  As faces are merged, the sign of
m_ID is flipped.

The sign is checked to identify valid flips.  This prevents double merges.

After the pass is complete, all faces that were never flipped (positive
m_ID) have their jref or kref halved.

ResetFaceNumbers() is called at the end to restore m_ID's primary
purpose.