OpenVSP 3.40.0
This version may not be huge, but it is certainly worthy.  It includes
a few small features, some quality of life improvements, API additions,
and a bunch of bug fixes.

Whether you'll be at AirVenture, Aviation, the OpenVSP Workshop, or none
of the above, this version should be exactly what you need for your
mid-summer modeling needs.

There are two cool additions to the propeller component.  Individual
blade folding control allows you to model stowed blades the way
rotorcraft typically do it rather than the more symmetrical approach
usually taken by folding propellers.  Also, there is now support for
non-uniorm blade spacing.  Non-uniorm spacing is a technique to reduce
the noise level of a prop / rotor.  In addition to allowing the user
to dial in the spacing, OpenVSP will automatically calculate the
positions of the final two blades to maintain rotor balance.

If you've been thinking that OpenVSP needs extensive online help -- and
you've been wishing there was a way you could contribute to OpenVSP as
a project -- have I got a deal for you.  OpenVSP now has the
infrastructure to make it incredibly easy to add a help screen to any
screen in the program.  There will be a presentation about this at the
Workshop, but if you want to give it a go before then, let me know and
I'll get you started.  Check out the new help on the Advanced Linking
screen as an example.

When switching to a multi-view mode, the default view selections were
not great.  Now the default 4-view mode matches the traditional choices
for an engineering drawing.  Some additional hot keys have been added
to make working with multiple views easier.  Try shift-F, shift-C, and
shift-R.

There are a bunch of API additions all around.  Some to improve automation
of graphics, materials, and screen shots.  Some to improve working with
FEA structures and generating meshes automatically.  Some to work with
the image files associated with 3D backgrounds.  It is great to see
what people are accomplishing with the API.

I've updated the build system to build packages for Ubuntu 24.04 -- and
I'm consequently dropping support for 20.04.  It should still work fine
if you need it, but you'll need to build for yourself.

There are a variety of bug fixes all around, but the one that the most
users will probably notice is that grid and tick lines have been
restored in the VSPAERO and Wave Drag plots.


Features:
 - Individual blade folding
 - Non-uniform blade spacing
 - Blade balance calculation and control
 - Online help infrastructure
 - Help cheat sheet for advanced linking
 - Improved 4-view view selection
 - Multi view control hot keys
 - FEA Mesh API additions
 - View and material setting API additions
 - 3D background file API additions

Build system:
 - Add support for Ubuntu 24.04 / drop 20.04
 - Update FLTK to latest 1.4.pre version

Bug fixes:
 - Fix grid lines not drawing on Windows
 - Adjust tolerance in MergeBorderEndPoints
 - Fix build system for facade generation
 - Use ParmContainer::EncodeXml/DecodeXml when writing simple links
 - Fix missing inputs for parasite drag analysis