Protect against invalid nodes in MatchBorderNodes.
For some reason, some members of nodeVec seemed to be valid pointers to
non-Node entities.  The best way to filter these out turned out to be
a try/catch block.  This prevented an outright crash and let things
generally keep working.

indexVec is needed so the indexing can be kept up with a now sparse
ptVec.