Move iQuad up a few levels to eliminate collisions.
TMeshes that were split because they are flat would end up with the iQuad
sequence re-starting.  This could cause erroneous merging in the extreme.

iQuad merging is also protected by Tag number, so moving this up to a
level that is guaranteed to have different tag numbers is sufficient to
protect against all erroneous merges.