OpenVSP 3.41.2
While there are some microfeatures in this version, it is really about
fixing some specific bugs in some relatively recently developed
capabilities.  As it gets put to use, some issues crop up, so here are
the fixes.

This version fixes a number of issues when using the GUI from the Facade
and the API.  If you are a plain GUI user or a non-GUI API user, you
shouldn't see any difference with this version (with some very minor
exceptions).

Fixes:
 - Improve displayed precision in User Parm sliders both int and float.
 - Add VSPCrash() to API to help test Wrappers.
 - Update on-screen file label when file name changes from GUI.
 - Make relative paths passed to SetVSP3ileName absolute.
 - Make vsp online help work from Python API, include files in package.
 - Improve finding python from facade.
 - Setup vspaero path from facade.
 - Make InitGUI() stand-alone and automatically call it.
 - Allow changes to GUI from API before it is open for first time.
 - Fix problems opening and closing GUI from API.