Clean up printing run flags to vspaero flag so OpenVSP can read them back
When OpenVSP reads in pre-existing VSPAERO solutions, it needs
to know some settings used when VSPAERO was run.  These can determine
what files to look for and how they should be parsed.

Some of these flags only appear on the command line and are therefore
not available -- so we write them to the *.vspaero file.  VSPAERO itself
ignores them.

We use the prefix VSP_ here to avoid any name collisions with VSPAERO
variables.