Use a flag to identify unset fpos_t
Formally, fpos_t is an 'opaque' type -- meaning the programmer can not
count on it having any underlying type or meaning -- and we can't set
it to zero or null to mean anything.

This worked fine on Clang and MSVC, but gcc balked.

On gcc, fpos_t is actually a struct with two values.