Add Step Children tracking to Geoms.
Ordinary children are descended from their parents (in terms of the Geom
tree) and depend on them according to attachment status or shape (for
conformal).

'Step' children depend on their parents, but are not descended from them.

This introduces a new kind of dependency -- one not established by
position in the family tree.  In particular, this will allow a single
Geom to be dependent on multiple parents (no longer their position in
the tree).

This will potentially introduce difficulties with circular dependency.