Rename 'Original' BBox to 'ScaleIndependent'.
Original was chosen when the only scale dependent geometry was the
inlet / exhaust extension for a GeomEngine.  It was the original
BBox because it was the BBox before the extension modifications were
performed.

Now that this mechanism is also used for landing gear ground plane
and clearance geometry keep-out zones, an update to the nomenclature
is appropriate.