Make GearGeom::UpdateBBox() consider ground plane location.
We don't want to consider the extents of the ground plane in the BBox as
the ground plane extents are determined by the size of the BBox and
an infinite loop of a growing ground plane will result.

However, ignoring the ground plane entirely results in strange
results when the model contains nothing but a simple landing gear
component.

Consequently, we consider the origin of the ground plane (with
symmetry applied via m_SymmTransMatVec)