Clean up 'ghost' surfaces in Geoms.
Long ago, m_MainSurfVec was initialized to non-empty in the Geom
constructor.  Ever since, there has been code that assumed there was
always at least one m_MainSurfVec.

Some Geoms overrided GetNumMainSurfs() to return zero -- even when
m_MainSurfVec was not empty.

This all goes back to when the main surface was a single member variable
insted of a vector.