Simplify Nerf/ManageGeomScreen to not cache vector of Geom pointers
The performance wins here were from looking up the Geom pointers using a
map.  Now that Vehicle::FindGeom(id) uses a map, we don't need to build
a map here.

The cache was only used in LoadBrowser(), so it does not make sense to
store it in the class and maintain it through Make/ClearGeomCache().
Instead, LoadBrowser can just do the equivalent operation once at the top.