Add ClearanceGeom::CalculateTurn()
This calculates the center of rotation and wheel track radii for a
three-wheeled landing gear configuration.

The steering is limited to the maximum available steering angle or the
ideal steering angle -- that which places the COR on one of the main
gear.  Using a greater steering angle would require the main wheels to
move in opposite directions, which usually requires a tug or other
complexities.