Fudge geomcopy# in vkey / groups file to support wings and props
This resets copyindex for any geom with only one main surface.  This is
usually the case.  Consequently, geomcopy# will follow geom# except when
there are multiple main surfaces (propellers).  This provides a numbering
where multiple blades of a prop are grouped together, but symmetrical
propellers are numbered separately.  However, symmetrical wings (and
other nmain==1 geoms) are numbered together.  This substitution needs
to be done every place we insert or lookup based on copyindex.

This is a bit of an ugly hack.  Instead of accessing the geom directly
for this information, it would be better if it was inferred from data
carried by the mesh itself.  However, it does not seem worth adding that
data and carrying it through for this.