BUGFIX veh-level user attribute duplication on file load
Vehicle::ReadXmlFile DecodesXml of ParmContainer, then DecodeXmlGeomsOnly
Both were reading vehicle-level attrs
AttributeCollection::DecodeXml checks if attributeID already registered or not when decoding the attribute
If the attribute already exists and is a non-random ID, then overwrite the old attribute with the read attributes' data
Otherwise, make a new attribute and add it to the collection
Logic hole was if an existing attributeID is being found and IS random, then it would make a duplicate which confuses OpenVSP and writes extra data to the xml file, exponentially growing the file size

Fix = 2 things
1. Remove the redundant vehicle-level attribute decode in the DecodeXmlGeomsOnly
That method only comes up when loading or inserting a vehicle file. When loading, it wypes first and then reads the veh attrs in the ParmContainer::DecodeXML; when inserting, it just inserts the "geoms" not vehicle data
2. For future proofing, come up with consistent logic for handling an ID crash when decoding the attribute at vehicle level.
a. Use old attr
b. Use new attr
c. ChangeID and add both

Current implementation opts for (a)
Adds bypass for unlikely case of decoding an attribute whose random ID is already in the
Includes a warning message to contact the devs about this harmless issue, but we should not run into it normally so it means something is weird with the attributes.

Revise decode error reporting to be less terrifying for users